Increase Meson warning level to 3
Fix all level-3 warnings across ah and parser-tests and enforce werror in CI.

Closes #45

+++ src/app/ahapp.cpp
@@ -393,7 +393,7 @@ void GetDirFromPath(const char * path, std::string & dir)
 
     dir = path;
     p   = dir.c_str() + (dir.size()-1);
-    while (*p!='\\' && *p!='/' && n<dir.size())
+    while (*p!='\\' && *p!='/' && static_cast<size_t>(n) < dir.size())
     {
         p--;
         n++;
